Move exceptions out of checkrunner megamodule into own module
